Threaing optimizations
git-svn-id: file:///Users/Eric/Developer/Other/YetiCodecSVN/trunk@86 a9bbc950-3a31-49f7-86fe-bb7552e34ce9

+++ convert_yuy2.h
@@ -1,5 +1,5 @@
 #pragma once
 #include "common.h"
 
-void mmx_ConvertRGB32toYUY2(const unsigned int* src, unsigned int* dst,int src_pitch, int dst_pitch, int w, int h);
+void mmx_ConvertRGB32toYUY2(const BYTE* src, BYTE* dst,int src_pitch, int dst_pitch, int w, int h);
 void mmx_ConvertRGB24toYUY2(const BYTE* src, BYTE* dst, int src_pitch, int dst_pitch, int w, int h);
